예약페이지 컴포넌트화
